requirements.txt
langchain
langchain-openai
langchain_ollama
python-dotenv
streamlit
langserve
fastapi
uvicorn
sse_starlette
ipykernel
langchain-community
beautifulsoup4
PyPDF
langchain-chroma
faiss-cpu
pymupdf
sentence-transformers
langgraph
grandalf
langchain_tavily
langchain-mcp-adapters
mcp
langgraph-cli[inmem]